Move `debug_collider_lines` from `core` to `scene` crate
