+++ game_editor/src/scene.rs
@@ -6,7 +6,6 @@ use game_common::components::{
 };
 use game_common::entity::EntityId;
 use game_common::world::{QueryWrapper, World};
-use game_core::debug::draw_collider_lines;
 use game_core::modules::Modules;
 use game_gizmos::Gizmos;
 use game_render::camera::{Camera, Projection, RenderTarget};
@@ -14,6 +13,7 @@ use game_render::entities::{CameraId, DirectionalLightId, PointLightId, SpotLigh
 use game_render::light::{DirectionalLight, PointLight, SpotLight};
 use game_render::scene::RendererScene;
 use game_render::Renderer;
+use game_scene::debug::draw_collider_lines;
 use game_scene::{InstanceId, SceneId, SceneSpawner};
 use game_tasks::TaskPool;
 use game_wasm::resource::ResourceId;
